rename shakeincubate package and split out anthaOS setup steps into seperate package
